+++ imports/email/email.js
@@ -1,5 +1,5 @@
 ThreadCategories.add("Email", {
   icon: "fa fa-envelope-open-o",
   iconUnread: "fa fa-envelope-o",
-  details: true
+  details: ['About', 'Members']
 });