+++ imports/roster/roster.js
@@ -1,7 +1,7 @@
 ThreadCategories.add("Roster", {
   icon: "fa fa-address-book-o fa-1-2x",
   iconUnread: "fa fa-address-book fa-1-2x",
-  details: true,
+  details: ['Members'],
   title(thread, detail=false) { // client only
     return I18n.t(thread.subject);
   },